v1.372 - Tooltip flotantes infinitos - El campo bono talento de los Skillstats es un float ahora en lugar de un Int, cambiado en todos los XML importadores de Datos - La bara de casteo solo se dibuja cuando se esta en cooldown, no cuando esta full - Creado Iterator en el Interfaz SpellI que devuelve el siguiente debuffQueAplica - Reescrito el codigo Drag and drop de mover las ventanas para que sea pixel perfect

+++ MyrranMP/MyrranCliente/src/DB/Datos/BDebuff/BDebuffLocalDB.java
@@ -76,7 +76,7 @@ public class BDebuffLocalDB
                     boolean isMejorable = Boolean.parseBoolean(stat.getChildText("isMejorable"));
                     int talentoMaximo   = Integer.parseInt(stat.getChildText("talentoMaximo"));
                     int costeTalento    = Integer.parseInt(stat.getChildText("costeTalento"));
-                    int bonoTalento     = Integer.parseInt(stat.getChildText("bonoTalento"));
+                    float bonoTalento   = Float.parseFloat(stat.getChildText("bonoTalento"));
 
                     debuff.skillStats()[numStat].setStat(nombreStat, valorBase);
                     if (isMejorable) debuff.skillStats()[numStat].setTalentos(talentoMaximo, costeTalento, bonoTalento);
